+++ examples/prime-sum-streaming.rs
@@ -6,6 +6,6 @@ fn main() {
     // now we can efficiently sum them up
     let sum = ns.iter()
                 .map(|n| primal::StreamingSieve::nth_prime(*n))
-                .fold(0, |a, b| a + b);
+                .fold(0u64, |a, b| a + b as u64);
     println!("the sum is {}", sum);
 }
